+++ app/routers/locations.py
@@ -15,8 +15,17 @@ from wwdtm.location import Location
 from app.config import API_VERSION, load_config
 from app.models.locations import Location as ModelsLocation
 from app.models.locations import LocationDetails as ModelsLocationDetails
+from app.models.locations import LocationID as ModelsLocationID
 from app.models.locations import Locations as ModelsLocations
 from app.models.locations import LocationsDetails as ModelsLocationsDetails
+from app.models.locations import LocationSlug as ModelsLocationSlug
+from app.models.locations import (
+    PostalAbbreviationDetails as ModelsPostalAbbreviationDetails,
+)
+from app.models.locations import PostalAbbreviations as ModelsPostalAbbreviations
+from app.models.locations import (
+    PostalAbbreviationsDetails as ModelsPostalAbbreviationsDetails,
+)
 
 router = APIRouter(prefix=f"/v{API_VERSION}/locations")
 _config = load_config()
@@ -208,8 +217,7 @@ async def get_location_recordings_by_id(
     except DatabaseError:
         raise HTTPException(
             status_code=500,
-            detail="Database error occurred while trying to "
-            "retrieve location information",
+            detail="Database error occurred while trying to retrieve location information",
         ) from None
 
 
@@ -251,6 +259,257 @@ async def get_location_recordings_by_slug(
     except DatabaseError:
         raise HTTPException(
             status_code=500,
-            detail="Database error occurred while trying to "
-            "retrieve location information",
+            detail="Database error occurred while trying to retrieve location information",
+        ) from None
+
+
+@router.get(
+    "/postal-abbreviations",
+    summary="Retrieve Postal Abbreviations",
+    response_model=ModelsPostalAbbreviations,
+    tags=["Locations"],
+)
+@router.head("/postal-abbreviations", include_in_schema=False)
+async def get_postal_abbreviations() -> list[str]:
+    """Retrieve All Postal Abbreviations.
+
+    Returned data: Postal abbreviations for states, provinces and
+    territoies.
+
+    Postal abbreviations are sorted alphabetically.
+    """
+    try:
+        location = Location(database_connection=_database_connection)
+        abbreviations = location.retrieve_postal_abbreviations()
+        if abbreviations:
+            return list(abbreviations.keys())
+
+        raise HTTPException(status_code=404, detail="No postal abbreviations found")
+    except ProgrammingError:
+        raise HTTPException(
+            status_code=500, detail="Unable to retrieve postal abbreviations"
+        ) from None
+    except DatabaseError:
+        raise HTTPException(
+            status_code=500,
+            detail="Database error occurred while trying to retrieve postal abbreviations",
+        ) from None
+
+
+@router.get(
+    "/postal-abbreviations/details",
+    summary="Retrieve Information for All Postal Abbreviations",
+    response_model=ModelsPostalAbbreviationsDetails,
+    tags=["Locations"],
+)
+@router.head("/postal-abbreviations/details", include_in_schema=False)
+async def get_postal_abbreviations_details() -> list[dict[str, str]]:
+    """Retrieve Details for all All Postal Abbreviations.
+
+    Returned data: Postal abbreviations for states, provinces and
+    territoies.
+
+    Postal abbreviations are sorted alphabetically.
+    """
+    try:
+        location = Location(database_connection=_database_connection)
+        abbreviations = location.retrieve_postal_abbreviations_list()
+        if abbreviations:
+            return {"postal_abbreviations": abbreviations}
+
+        raise HTTPException(status_code=404, detail="No postal abbreviations found")
+    except ProgrammingError:
+        raise HTTPException(
+            status_code=500, detail="Unable to retrieve postal abbreviations"
+        ) from None
+    except DatabaseError:
+        raise HTTPException(
+            status_code=500,
+            detail="Database error occurred while trying to retrieve postal abbreviations",
+        ) from None
+
+
+@router.get(
+    "/postal-abbreviations/details/{abbreviation}",
+    summary="Retrieve Information for a Postal Abbreviation",
+    response_model=ModelsPostalAbbreviationDetails,
+    tags=["Locations"],
+)
+@router.head("/postal-abbreviations/details/{abbreviation}", include_in_schema=False)
+async def get_postal_abbreviation(
+    abbreviation: Annotated[
+        str, Path(title="Postal Abbreviation to retrieve information")
+    ],
+):
+    """Retrieve Details for a Postal Abbreviation.
+
+    Returned data: Postal abbreviation, name of the state, province or
+    territory, and country.
+    """
+    try:
+        location = Location(database_connection=_database_connection)
+        info = location.retrieve_postal_details_by_abbreviation(
+            abbreviation=abbreviation
+        )
+        if info:
+            return info
+
+        raise HTTPException(
+            status_code=404,
+            detail=f"Postal abbreviation {abbreviation} not found",
+        )
+    except ValueError:
+        raise HTTPException(
+            status_code=404,
+            detail=f"Postal abbreviation {abbreviation} not found",
+        ) from None
+    except ProgrammingError:
+        raise HTTPException(
+            status_code=500, detail="Unable to retrieve postal abbreviation information"
+        ) from None
+    except DatabaseError:
+        raise HTTPException(
+            status_code=500,
+            detail="Database error occurred while trying to retrieve postal "
+            "abbreviaton information",
+        ) from None
+
+
+@router.get(
+    "/random",
+    summary="Retrieve Information for a Random Location",
+    response_model=ModelsLocation,
+    tags=["Locations"],
+)
+@router.head("/random", include_in_schema=False)
+async def get_random_location():
+    """Retrieve a Random Location.
+
+    Returned data: Location ID, venue, city, state and slug string.
+    """
+    try:
+        location = Location(database_connection=_database_connection)
+        location_info = location.retrieve_random()
+        if location_info:
+            return location_info
+
+        raise HTTPException(status_code=404, detail="Random Location not found")
+    except ValueError:
+        raise HTTPException(
+            status_code=404, detail="Random Location not found"
+        ) from None
+    except ProgrammingError:
+        raise HTTPException(
+            status_code=500, detail="Unable to retrieve location information"
+        ) from None
+    except DatabaseError:
+        raise HTTPException(
+            status_code=500,
+            detail="Database error occurred while trying to retrieve location information",
+        ) from None
+
+
+@router.get(
+    "/random/details",
+    summary="Retrieve Information and Appearances for a Random Location",
+    response_model=ModelsLocationDetails,
+    tags=["Locations"],
+)
+@router.head("/random/details", include_in_schema=False)
+async def get_random_location_details():
+    """Retrieve a Random Location.
+
+    Returned data: Location ID, venue, city, state, slug string, and recordings.
+
+    Appearances are sorted by date.
+    """
+    try:
+        location = Location(database_connection=_database_connection)
+        location_details = location.retrieve_random_details()
+        if location_details:
+            return location_details
+
+        raise HTTPException(status_code=404, detail="Random Location not found")
+    except ValueError:
+        raise HTTPException(
+            status_code=404, detail="Random Location not found"
+        ) from None
+    except ProgrammingError:
+        raise HTTPException(
+            status_code=500, detail="Unable to retrieve location information"
+        ) from None
+    except DatabaseError:
+        raise HTTPException(
+            status_code=500,
+            detail="Database error occurred while trying to retrieve location information",
+        ) from None
+
+
+@router.get(
+    "/random/id",
+    summary="Retrieve a Random Location ID",
+    response_model=ModelsLocationID,
+    tags=["Locations"],
+)
+@router.head("/random/id", include_in_schema=False)
+async def get_random_location_id():
+    """Retrieve a Random Location ID.
+
+    Returned data: Location ID.
+    """
+    try:
+        location = Location(database_connection=_database_connection)
+        location_id = location.retrieve_random_id()
+        if location_id:
+            return {"id": location_id}
+
+        raise HTTPException(status_code=404, detail="Random Location ID not returned")
+    except ValueError:
+        raise HTTPException(
+            status_code=404, detail="Random Location ID not returned"
+        ) from None
+    except ProgrammingError:
+        raise HTTPException(
+            status_code=500, detail="Unable to retrieve a random location ID"
+        ) from None
+    except DatabaseError:
+        raise HTTPException(
+            status_code=500,
+            detail="Database error occurred while trying to retrieve a random location ID",
+        ) from None
+
+
+@router.get(
+    "/random/slug",
+    summary="Retrieve a Random Location Slug String",
+    response_model=ModelsLocationSlug,
+    tags=["Locations"],
+)
+@router.head("/random/slug", include_in_schema=False)
+async def get_random_location_slug():
+    """Retrieve a Random Location Slug String.
+
+    Returned data: Location slug string.
+    """
+    try:
+        location = Location(database_connection=_database_connection)
+        location_slug = location.retrieve_random_slug()
+        if location_slug:
+            return {"slug": location_slug}
+
+        raise HTTPException(
+            status_code=404, detail="Random Location slug string not returned"
+        )
+    except ValueError:
+        raise HTTPException(
+            status_code=404, detail="Random Location slug string not returned"
+        ) from None
+    except ProgrammingError:
+        raise HTTPException(
+            status_code=500, detail="Unable to retrieve a random location slug string"
+        ) from None
+    except DatabaseError:
+        raise HTTPException(
+            status_code=500,
+            detail="Database error occurred while trying to retrieve a location host slug string",
         ) from None